requirements.txt
requests
networkx
matplotlib
beautifulsoup4
python-dotenv
